opensync: dont use the @wan syntax
The @ syntax reference the L3 device. This means a child interface will not
be brought up if the underlying parent is still pending.

This can be seen if we have br-wan and br-wan.100. Both have proto dhcp.
The vlan interface wont come up until the wan interface has an IP.

Signed-off-by: John Crispin <john@phrozen.org>