opensync: add kernel crash detection
This feature makes use of the crashlog driver in the kernel. If we experience
an oops, the kernel will store the stack trace inside /sys/kernel/debug/crashlog.
OpenSync will then add an entry to the Alarms table. The could can then retrieve
the crashlog via the Command_config table.

A crash can be simulated with
   echo c > /proc/sysrq-trigger

Signed-off-by: John Crispin <john@phrozen.org>