ath10k-ct:  Don't override regulatory domain.
A painfully slow bug hunt lead to the fact that edge-core 5410-L EU
variant has radios that hang when the driver sets the regulatory domain
to USA.

So, stop overriding regulatory domain by default.  If a user wants to
do this on a platform, it can be changed on their specific platform.

While at it, increase tx buffers for 9984 a bit since we have some
extra RAM to spare.  This will help scale high tx speed with lots of
stations associated since there is more buffer space to build larger ampdus.

Signed-off-by: Ben Greear <greearb@candelatech.com>