apc: Add Access Point coordinator choosing algorithm
Add algorithm which chooses a APC and
backup APC among the APs in a subnet.

This APC can be used to perform few functions
on behalf of the APs on the subnet. For example,
the APC AP will be used as a radius proxy for
the APs in the subnet.

Signed-off-by: Chaitanya Godavarthi <chaitanya.kiran@netexperience.com>