ucentral: Don't use batmand
batman-adv is a layer 2 implementation of the B.A.T.M.A.N. (IV/V) algorithm.
The batmand is an old layer 3 implementation of the B.A.T.M.A.N. III
algorithm. The latter is not really useful when already using batman-adv on
top of the IBSS/meshpoint interfaces. And running layer 3 B.A.T.M.A.N. III
on top of layer 2 B.A.T.M.A.N. IV might lead to all kind of unexpected
behavior.

Signed-off-by: Sven Eckelmann <sven@narfation.org>