wifi-4421: Fix client disconn when in DFS channel
On detecting Radar on the DFS channel the radio switches
the channel to backup channel. However, the opensync tries
to set it back to the configured channel every 30 seconds, which
which would fail since the channel is in a radar NOP period, this
leads to dropping of clients.
Fix by adding a check for the configured channel's availability
before trying to change to that channel.
Also, check if the radio Channel config has changed before
attempting to change the channel.

Signed-off-by: Chaitanya Godavarthi <chaitanya.kiran@netexperience.com>