Wifi-5798. Fix for SM crash during VoIP calls
nf_uccdetect module tries and analyze the traffic to detect
VoIP calls, and generate events up to the user space. This module
assigns value -1 as RadioID for packets from Ethernet interface
which eventually is used as an array index, eventually running
into an Assert and crash. Fix is to skip these packets from
Ethernet interface for the analysis.

Signed-off-by: ravi vaishnav <ravi.vaishnav@netexperience.com>