minim: Build unum via top level Makefile
Before this commit, much of the logic in this feed makefile
was copied from the top level unum-sdk makefile, and the unum
agent was built by calling make in src/unum.  Copying of code is
usually not a good thing from a maintenance perspective and by not
leveraging the existing unum-sdk build process the feed makefile needed
to have knowledge of what to install from the build output and source
tree.

Now, call the top level unum-sdk makefile to build the agent and let
the install rules in the unum-sdk source decide what gets installed
where.  Then, simply install the contents of the resulting unum rootfs
tarball as the unum package.

SW-815