uspot: further abstract from spotfilter
This commit further detaches the portal frontend from spotfilter by
interfacing with the uspot.uc backend instead of querying spotfilter
directly.

The backend maintains its own internal state independently of
spotfilter, and exchanges a curated view of the client state with the
frontend, thus limiting potential secret leaks.

For debugging purposes, when debug is enabled on a given interface, the
entire internal client state is saved in spotfilter 'data'.

Signed-off-by: Thibaut VARÈNE <hacks@slashdirt.org>