uspot: work around ucode divergence
In ucode, the '?.' operator may not short-circuit when using square
brackets. Quoting the proposed fix:

<jow> for now you can use ?. (or `?.[` / `?.(`) throughout the entire
chain, it will be forward compatible when/if short circuiting lands

In client_add(), with the support of additional authentication methods,
the radius property may not always exist, causing a crash as the
short-circuit is not taken.

Signed-off-by: Thibaut VARÈNE <hacks@slashdirt.org>