WIFI-13797:Fix Modifying VLAN-ID Under LAN Port : using event handling
Fixes WIFI-13797: This commit is an improvement over previous
commit 3598a243b31f4276cbe4e54eecb19281c5d2f884 which added a
fix for traffic disruption when vlan id of the lan port is
changed on devices with internal switches such as fap655,
CIG_WF186w,EAP-104.
This commit reconfigures the vlans for the mpsk clients on the
switch using the ucentral-event subsystem by tracking the
netifd_add events.

Signed-off-by: joydeepbenison <joydeep.ghosh@benisontech.com>